+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Patch Notes
 
+## Version 1.0.1
+
+* Add fullscreen button.
+* Fix eshint version detection.
+* Add line wrapping.
+
 ## Version 1.0.0
 
 * Initial release.